grep_log/requirements.txt
pyahocorasick==1.4.0